+++ custom_components/aqara_m1s_zigbee_router/const.py
@@ -25,6 +25,7 @@ SERVICE_DELETE_SOUND = "delete_sound"
 SERVICE_REFRESH_SOUNDS = "refresh_sounds"
 SERVICE_RESET_MEDIA_GROUP = "reset_media_group"
 SERVICE_RESYNC_MEDIA_GROUP = "resync_media_group"
+SERVICE_UPDATE_MEDIA_METADATA = "update_media_metadata"
 
 SOUND_ROOT = "/data/musics"
 MANAGED_SOUND_ROOT = "/data/musics/music-ch"